pyproject.toml
[project]
name = "mineral-rdf-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.1",
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "langgraph-supervisor>=0.0.26",
    "langsmith>=0.3.42",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.5",
    "pytesseract>=0.3.13",
]
